backend/requirements.txt
Django==4.2.9
HAP-python==4.9.1
aiohttp==3.8.5
cryptography==41.0.2
daphne==4.0.0
dj-stripe==2.8.3
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-filter==23.2
django-formtools==2.4.1
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-redis==5.4.0
django-storages==1.13.2
django-timezone-field==5.0
django-versatileimagefield==3.0
djangorestframework-simplejwt==5.3.0
djangorestframework==3.14.0
gunicorn==21.2.0
huesdk==1.7
ipython==8.20.0
mssql-django==1.3
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.0
sqlparse==0.4.3
zeroconf==0.131.0
psutil==5.9.8
scapy==2.5.0